Remove action map item and ProblemListActivity
